Add files via upload
What is the option value of the right to buy shares in a company on a future date at the current price? The code simulates the value of the stock on that future date based on randomly selected observed daily price % changes, then discounts the difference in that future value vs. the strike price when greater than 0 to calculate a present value of the option as a mean and shows the %ile spread of simulated outcomes.