Add files via upload
What is the probable distribution of price for a given equity in 4 years? For applications such as option pricing and portfolio risk modeling, we want to understand more than expected return. In other words, shape matters.

This example fits % price changes in the S&P 500 to a gamma distribution, then applies that fit to simulate the value of a $100 investment in the S&P 500
in 4 years. A parameterized Monte Carlo approach allows for exploration of a. how parameters shift with different time slices of data, b. comparison of parameters across other assets and asset classes, and c. parameter blending with other similar assets to improve predictive power and model credibility.